% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashpipe.R
\name{hashpipe}
\alias{hashpipe}
\title{hashpipe
This function sends '#| ' at the current cursor position. You'd never run this function using its name! Use it as an add-in for a keyboard shortcut.}
\usage{
hashpipe()
}
\value{
string with '#| ' minus the quote marks
}
\description{
hashpipe
This function sends '#| ' at the current cursor position. You'd never run this function using its name! Use it as an add-in for a keyboard shortcut.
}
